+++ NEWS.md
@@ -1,3 +1,11 @@
+# ggfittext (development version)
+
+## Minor changes
+
+- `contrast = TRUE` now assumes ggplot2 default colours for `fill` and
+  `panel.background`, so it can be used in the absence of a `fill` aesthetic.
+- `geom_bar_text()` now defaults to `contrast = TRUE`.
+
 # ggfittext 0.8.1
 
 ## Minor changes